feat: WebUI TLS fixes, IRCv3 batch/EXTBAN, SASL SCRAM, libop primitives
Bug fixes:
- tls.c: inverted opssl error checks (!=0 vs ==0) caused cert/key load
  to fail silently or fatal on success; WebUI TLS was broken
- net.c: direction-aware TLS handshake — always called opssl_connect()
  even for incoming connections; now uses SOCK_CONNECT to pick connect/accept
- net.c: write-driven TLS handshake in dequeue_sockets for WANT_WRITE resume;
  arm write events for all mid-handshake TLS sockets
- userrec.c: SIGSEGV in get_user_by_host / del_host_or_account — USERENTRY_HOSTS
  is op_vec_t* but was iterated as struct list_type*; crash on any connection
- eggdrop-basic.toml: WebUI listener now uses +8443 (TLS prefix)

IRC/server features:
- IRCv3 BATCH accumulator with generic H_batch bind table; chathistory compat
- current_msgtag_account and current_msgtag_batch exported for handler use
- EXTBAN= ISUPPORT parsed; extban_match() for extended ban checking
- CAP REQ queue via server_queue_cap_req()
- SASL: SCRAM-SHA-1 mechanism; per-username backoff rate limiting; TLS guard
- Dynamic ISUPPORT limits (MODES=, CHANLIMIT=, CHANNELLEN=, PREFIX=) live-updated
- New Tcl: chathistory_latest, curserver, getcurrentnick, serverport, serverlag,
  isupportval, isupportprefix, chanstats, chaninfo
- New IRC numeric handlers: 301, 312-313, 319, 321-323, 333-334, 341, 351, 391

WebUI:
- Per-IP auth rate limiting with exponential lockout window
- webui:nick_push Tcl command; real-time TOPIC push to WebUI clients

Core:
- match.c: wildcard patterns compiled to ERE via irc_to_ere + op_regex_match
- userrec.c: heap alignment fixes; mark_user_dirty/mark_all_users_dirty visibility

libop:
- New primitives: op_atomic, op_lock, op_memory, op_mpmc (MPMC queue),
  op_thread_pool, op_event (unified poll abstraction)
- commio.c: edge/level-trigger fix, per-connection state machine, zero-copy read
- New utilities: hostmask, msgbuf, regex, subst

opssl:
- TLS 1.3 handshake correctness; DTLS driver; ML-KEM hybrid coverage

+++ subprojects/opssl/src/x509/asn1.c
@@ -420,7 +420,9 @@ opssl_asn1_get_time(opssl_cbs_t *cbs, int64_t *epoch)
 
     days += day - 1; /* Zero-based day */
 
-    *epoch = days * 86400 + hour * 3600 + min * 60 + sec;
+    /* Promote to int64_t before multiplication to avoid 32-bit overflow. */
+    *epoch = days * (int64_t)86400 + (int64_t)hour * 3600 +
+             (int64_t)min * 60 + (int64_t)sec;
     return 1;
 }
